Rename FailedGarbageCollectionException
